refactor: api type 수정 및 폴더 types로 이동
- api/types에서 types/api로 폴더 이동
- api type 백엔드 data에 맞춰서 변경 (빠진 데이터 주석으로 메모)
- 도메인으로 파일 분류, 합칠 수 있는 것들은 하나의 파일로 그룹핑
