--- api/types/error.ts
@@ -1,4 +1,0 @@
-export interface IErrorResponse {
-  status: number;
-  message: string;
-}
