--- api/types/menuList.ts
@@ -1,7 +1,0 @@
-export interface IMenuItem {
-  id: number;
-  name: string;
-  menuImage: string;
-  summary: string;
-  price: number;
-}
