+++ types/api/axios.ts
@@ -1,0 +1,12 @@
+import { AxiosResponse } from "axios";
+
+export interface IErrorResponse {
+  status: number;
+  message: string;
+}
+
+export interface IBaseRepsonse<T> extends AxiosResponse<T> {
+  status: number;
+  data: T;
+  message: string;
+}
